Added prompt for overwriting files

+++ build/lib/copy_to/__init__.py
@@ -278,6 +278,10 @@ def copy_to(dest, src):
             else:
                 raise SystemExit
         exist_dest=os.path.join(dest, os.path.basename(os.path.normpath(element)))
+        if os.path.isfile(exist_dest):
+            res = prompt("There's already a file in the destination: " + exist_dest + ". Should it be overwritten? [y/n]: ", pre_run=prompt_autocomplete, completer=WordCompleter(["y", "n"]))
+            if res == "n":
+                continue
         if os.path.isfile(element):
             shutil.copy2(element, exist_dest)
             print("Copied to " + str(exist_dest))
@@ -295,6 +299,10 @@ def copy_from(dest, src):
             else:
                 raise SystemExit
         exist_dest=os.path.join(dest, os.path.basename(os.path.normpath(element)))
+        if os.path.isfile(exist_dest):
+            res = prompt("There's already a file in the destination: " + exist_dest + ". Should it be overwritten? [y/n]: ", pre_run=prompt_autocomplete, completer=WordCompleter(["y", "n"]))
+            if res == "n":
+                continue
         if os.path.isfile(exist_dest):
             shutil.copy2(exist_dest, element)
             print("Copied to " + str(element))
@@ -484,18 +492,22 @@ def get_group_subparsers(subparser):
     delete_from_group = subparser.add_parser('delete-from-group')
     add_group_parser = add_to_group.add_subparsers(dest='group')
     delete_group_parser = delete_from_group.add_subparsers(dest='group')
-    names = {}
+    names_add = {}
+    names_del = {}
     for name, value in conf.envs.items():
         if name == 'group':
-            add_parser = add_group_parser.add_parser(name)
-            delete_parser = delete_group_parser.add_parser(name)
-            names[name]=[]
             for i in value:
+                add_parser = add_group_parser.add_parser(i)
+                delete_parser = delete_group_parser.add_parser(i)
+                names_add[i]=[]
+                names_del[i]=[]
                 for e in get_reg_names():
                     if not e in conf.envs['group'][i]:
-                        names[name].append(e)
-            add_parser.add_argument("src" , nargs='+', help="Source number/Source Number Range", metavar="Source number/Source Number Range", choices=names[name])
-            delete_parser.add_argument("src" , nargs='+', help="Source number/Source Number Range", metavar="Source number/Source Number Range", choices=names[name])
+                        names_add[i].append(e)
+                    elif e in conf.envs['group'][i]:
+                        names_del[i].append(e)
+                add_parser.add_argument("name" , nargs='+', help="Group name holding multiple configuration names", metavar="Group name", choices=names_add[i])
+                delete_parser.add_argument("name" , nargs='+', help="Configuration name", metavar="Configuration name(s)", choices=names_del[i])
     return add_to_group,delete_from_group
 
 
@@ -619,6 +631,7 @@ def ask_git(prmpt="Setup git configuration to copy objects between? [y/n]: "):
         res_path = pathlib.Path(res)
         if git_path in res_path.parents:
             res = str(res_path.relative_to(os.path.dirname(git.Repo("./", search_parent_directories=True).git_dir)))
+            res = os.path.realpath(res)
         with repo.config_writer() as confw:
             confw.set_value("copy-to", "file", res)
         print("Added file = " + str(res) + " to local git configuration")
@@ -626,6 +639,10 @@ def ask_git(prmpt="Setup git configuration to copy objects between? [y/n]: "):
         with repo.config_writer() as confw:
             confw.set_value("copy-to", "run", res)
         print("Added run = " + str(res) + " to local git configuration")
+        res = prompt("Overwrite existing files?: (prevents prompt - y(es) / n(o)): ", pre_run=prompt_autocomplete, completer=WordCompleter(["y", "n"]))
+        with repo.config_writer() as confw:
+            confw.set_value("copy-to", "overwrite", res)
+        print("Added overwrite = " + str(res) + " to local git configuration")
         return res
     else:
         res1 = prompt("You selected no. Prevent this popup from coming up again? [y/n]:", pre_run=prompt_autocomplete, completer=WordCompleter(["y", "n"]))
@@ -648,7 +665,7 @@ def main():
     delete = subparser.add_parser('delete')
     add_source = subparser.add_parser('add-source')
     set_source_path = get_sourcepath_subparsers(subparser)
-    add_to_group,delete_from_group = get_group_subparsers(subparser)
+    add_to_group, delete_from_group = get_group_subparsers(subparser)
     if shutil.which("git"):
         set_git = subparser.add_parser('set-git')
     add_group = subparser.add_parser('add-group')
@@ -676,18 +693,21 @@ def main():
 
     delete_group.add_argument("groupname" , type=lambda x: is_valid_group(parser, x) ,help="Group name holding multiple configuration names", metavar="Group name", choices=get_group_names())
     
-    add_to_group.add_argument("groupname", type=lambda x: is_valid_group(parser, x), help="Group name holding multiple configuration names", metavar="Group name", choices=get_group_names())
-    add_to_group.add_argument("name" , nargs='+', type=lambda x: is_valid_name(parser, x), help="Configuration name", metavar="Configuration name", choices=get_reg_names())
+    #add_to_group.add_argument("groupname", type=lambda x: is_valid_group(parser, x), help="Group name holding multiple configuration names", metavar="Group name", choices=get_group_names())
+    #add_to_group.add_argument("name" , nargs='+', type=lambda x: is_valid_name(parser, x), help="Configuration name", metavar="Configuration name", choices=get_reg_names())
 
-    delete_from_group.add_argument("groupname" , type=lambda x: is_valid_group(parser, x), help="Group name holding multiple configuration names", metavar="Group name", choices=get_group_names())
-    delete_from_group.add_argument("name" , nargs='+', type=lambda x: is_valid_name(parser, x), help="Configuration name", metavar="Configuration name", choices=get_reg_names())
+    #delete_from_group.add_argument("groupname" , type=lambda x: is_valid_group(parser, x), help="Group name holding multiple configuration names", metavar="Group name", choices=get_group_names())
+    #delete_from_group.add_argument("name" , nargs='+', type=lambda x: is_valid_name(parser, x), help="Configuration name", metavar="Configuration name", choices=get_reg_names())
     
     if shutil.which("git"):
         git_command = set_git.add_subparsers(dest='gitcommand')
         git_run = git_command.add_parser('run')
         git_run.add_argument("value" , nargs='?' ,type=str , help="Configuration name", metavar="Configuration name", choices=get_names(True))
         git_file = git_command.add_parser('file')
-        git_file.add_argument("value" , nargs='?' ,type=lambda x: is_valid_conf(parser, x) , help="Configuration file", metavar="Configuration file")
+        git_file.add_argument("value" , nargs='1' ,type=lambda x: is_valid_conf(parser, x) , help="Configuration file", metavar="Configuration file")
+        git_overwrite = git_command.add_parser('overwrite')
+        git_overwrite.add_argument("value" , nargs='1' ,type=str , help="Overwrite files?", metavar="Configuration file", choices=['y','n'])
+        
     
     add_source.add_argument("name" , type=str ,help="Configuration name for modifications", metavar="Configuration name",  choices=get_reg_names())
     add_source.add_argument("src" , nargs='+', type=lambda x: is_valid_file_or_dir(parser, x), metavar="Source files and directories", help="Source files and directories")
@@ -797,8 +817,8 @@ def main():
                 listName(name)
     elif args.command == 'set-git':
         if not args.gitcommand:
-            args.gitcommand = prompt("Give up a git value to set (run/file): ", pre_run=prompt_autocomplete, completer=WordCompleter(["run", "file"]))
-            
+            args.gitcommand = prompt("Give up a git value to set (run/file/overwrite): ", pre_run=prompt_autocomplete, completer=WordCompleter(["run", "file", 'overwrite']))
+
         if args.gitcommand == 'run':
            repo = git.Repo("./", search_parent_directories=True)
            res = "all"
@@ -848,11 +868,14 @@ def main():
            res_path = pathlib.Path(res)
            if git_path in res_path.parents:
                res = str(res_path.relative_to(os.path.dirname(git.Repo("./", search_parent_directories=True).git_dir)))
-               #res = os.path.relpath(res)
+               res = os.path.realpath(res)
            with repo.config_writer() as confw:
                confw.set_value("copy-to", "file", res)
            print("Added " + str(res) + " to local git configuration")
-        
+       elif args.gitcommand == 'overwrite':
+           repo = git.Repo("./", search_parent_directories=True)
+           with repo.config_writer() as confw:
+               confw.set_value("copy-to", "run", args.value)
 
     elif args.command == 'add':
         if not 'name' in args:
@@ -939,29 +962,29 @@ def main():
                 json.dump(conf.envs, outfile)
 
     elif args.command == 'add-to-group':
-        if not 'groupname' in args:
+        if not 'group' in args:
             print("Give up an configuration to copy objects between")
             raise SystemExit
-        elif not args.groupname in conf.envs['group']:
-            print(str(args.groupname) + ". Doesn't exist. Look again.")
+        elif not args.group in conf.envs['group']:
+            print(str(args.group) + ". Doesn't exist. Look again.")
             listGroupNames()
             raise SystemExit
         else:
             groupnames = []
+            groupnames = conf.envs['group'][args.group]
+            for known in groupnames:
+                for n in args.group:
+                    if known == n:
+                        print("Look again. " + known + " is already in " + args.group)
+                        listAllGroups()
+                        raise SystemExit
             with open(conf.file, 'w') as outfile: 
-                groupnames = conf.envs['group'][args.groupname]
-                for known in groupnames:
-                    for n in args.name:
-                        if known == n:
-                            print("Look again. " + known + " is already in " + args.groupname)
-                            listAllGroups()
-                            raise SystemExit
                 for key in name:
-                    conf.envs['group'][args.groupname].append(key)
+                    conf.envs['group'][args.group].append(key)
                 json.dump(conf.envs, outfile)
             if args.list:
-                    listName(args.groupname)
-            print(str(args.name) + ' added to ' + str(args.groupname))
+                listName(args.group)
+            print(str(args.name) + ' added to ' + str(args.group))
 
     elif args.command == 'delete':
         if not 'name' in args:
@@ -1007,29 +1030,29 @@ def main():
                 json.dump(conf.envs, outfile)
     
     elif args.command == 'delete-from-group':
-        if not 'groupname' in args:
+        if not 'group' in args:
             print("Give up an configuration to copy objects between")
             raise SystemExit
-        elif not args.groupname in conf.envs['group']:
-            print(str(args.groupname) + ". Doesn't exist. Look again.")
+        elif not args.group in conf.envs['group']:
+            print(str(args.group) + ". Doesn't exist. Look again.")
             listGroupNames()
             raise SystemExit
         else:
-            groupnames = []
-            with open(conf.file, 'w') as outfile: 
-                groupnames = conf.envs['group'][args.groupname]
-                for known in groupnames:
-                    for n in args.name:
-                        if known == n:
-                            conf.envs['group'][args.groupname].remove(known)
-                            args.name.remove(n)
-                            if args.list:
-                                listName(args.groupname) 
-                            print(str(n) + ' removed from ' + str(args.groupname))
-                json.dump(conf.envs, outfile)
-                if not args.name == []:
-                    print("Look again. " + str(args.name) + " isn't/aren't in " + str(args.groupname))
-                    raise SystemExit
+            group = []
+            groupnames = conf.envs['group'][args.group]
+            for known in conf.envs['group'][args.group]:
+                for n in args.name:
+                    if known == n:
+                        with open(conf.file, 'w') as outfile: 
+                            conf.envs['group'][args.group].remove(known)
+                            json.dump(conf.envs, outfile)
+                        args.name.remove(n)
+                        if args.list:
+                            listName(args.group) 
+                        print(str(n) + ' removed from ' + str(args.group))
+            if not args.name == []:
+                print("Look again. " + str(args.name) + " isn't/aren't in " + str(args.group))
+                raise SystemExit
     elif args.command == 'add-source':
         if not 'name' in args:
             print("Give up an configuration to copy objects between")
